requirements.txt
PyGithub>=2.1.1
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.8.0
markdown>=3.5.0
requests>=2.28.0
